+++ piafedit/app.py
@@ -1,10 +1,6 @@
 import logging
-import sys
 
-import pyqtgraph as pg
-import pyqttango
 from PyQt5.QtGui import QIcon
-from PyQt5.QtWidgets import QApplication
 
 from piafedit.editor_api import P
 from piafedit.gui.main_ui import MainUi
@@ -12,7 +8,7 @@ from piafedit.model.geometry.size import SizeAbs
 from piafedit.model.source.rio_data_source import RIODataSource
 from piafedit.model.source.window import Window
 from piafedit.model.work_model import WorkModel
-from piafedit.ui_utils import resources_path
+from piafedit.ui_utils import resources_path, gui_app
 
 IMAGE_SIZE = 10_000
 
@@ -33,13 +29,10 @@ def gen_big_image():
 
 
 def main():
-    pyqttango.init_resources()
-    model = WorkModel(resources_path())
-    pg.setConfigOptions(imageAxisOrder='row-major')
-    app = QApplication([])
-    QIcon.setThemeName('tango')
-    P.main_window = MainUi(model)
-    sys.exit(app.exec_())
+    with gui_app():
+        model = WorkModel(resources_path())
+        QIcon.setThemeName('tango')
+        P.main_window = MainUi(model)
 
 
 if __name__ == '__main__':