+++ tests/qt_test/test_editor.py
@@ -1,16 +1,13 @@
-import sys
-
-from PyQt5 import QtWidgets
 from qutepart import Qutepart
 
+from piafedit.ui_utils import gui_app
+
 if __name__ == '__main__':
-    app = QtWidgets.QApplication([])
-    editor = Qutepart()
-    editor.completionEnabled = True
-    with open(__file__) as _:
-        editor.text = _.read()
-    editor.show()
+    with gui_app():
+        editor = Qutepart()
+        editor.completionEnabled = True
+        with open(__file__) as _:
+            editor.text = _.read()
+        editor.show()
 
-    editor.detectSyntax(xmlFileName='python.xml')
-
-    sys.exit(app.exec_())
+        editor.detectSyntax(xmlFileName='python.xml')