+++ tests/qt_test/test_palette.py
@@ -1,12 +1,8 @@
-import sys
-
-from PyQt5 import QtWidgets
-from PyQt5.QtGui import QPalette
 from PyQt5.QtWidgets import QColorDialog
 
-if __name__ == "__main__":
-    app = QtWidgets.QApplication([])
-    window = QColorDialog()
+from piafedit.ui_utils import gui_app
 
-    window.show()
-    sys.exit(app.exec_())
+if __name__ == "__main__":
+    with gui_app():
+        window = QColorDialog()
+        window.show()