+++ tests/qt_test/test_tango_qt.py
@@ -1,16 +1,12 @@
-import sys
-
-import pyqttango
 from PyQt5 import QtGui
 from PyQt5.QtCore import QSize
-from PyQt5.QtWidgets import QPushButton, QApplication
+from PyQt5.QtWidgets import QPushButton
+
+from piafedit.ui_utils import gui_app
 
 if __name__ == "__main__":
-    pyqttango.init_resources()
-    app = QApplication([])
-    QtGui.QIcon.setThemeName('tango')
-    win = QPushButton()
-    win.setIconSize(QSize(128, 128))
-    win.setIcon(QtGui.QIcon.fromTheme('accessories-calculator'))
-    win.show()
-    sys.exit(app.exec_())
+    with gui_app():
+        win = QPushButton()
+        win.setIconSize(QSize(128, 128))
+        win.setIcon(QtGui.QIcon.fromTheme('accessories-calculator'))
+        win.show()